add MVVM architecture
