+++ app/src/main/java/ru/musindev/myapp/domain/Interactor.kt
@@ -1,0 +1,7 @@
+package ru.musindev.myapp.domain
+
+import ru.musindev.myapp.data.MainRepository
+
+class Interactor(val repo: MainRepository) {
+    fun getFilmsDB(): List<Film> = repo.filmsDataBase
+}
