+++ app/src/main/java/ru/musindev/myapp/view/MainActivity.kt
@@ -1,4 +1,4 @@
-package ru.musindev.myapp
+package ru.musindev.myapp.view
 
 import android.os.Bundle
 import android.view.Window
@@ -6,7 +6,14 @@ import android.view.WindowManager
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.fragment.app.Fragment
+import ru.musindev.myapp.R
 import ru.musindev.myapp.databinding.ActivityMainBinding
+import ru.musindev.myapp.domain.Film
+import ru.musindev.myapp.view.fragments.DetailsFragment
+import ru.musindev.myapp.view.fragments.FavoritesFragment
+import ru.musindev.myapp.view.fragments.HomeFragment
+import ru.musindev.myapp.view.fragments.SelectionsFragment
+import ru.musindev.myapp.view.fragments.WatchLaterFragment
 
 @Suppress("DEPRECATION")
 class MainActivity : AppCompatActivity() {
