+++ app/src/main/java/ru/musindev/myapp/view/customviews/RatingDonutView.kt
@@ -1,4 +1,4 @@
-package ru.musindev.myapp
+package ru.musindev.myapp.view.customviews
 
 import android.annotation.SuppressLint
 import android.content.Context
@@ -9,11 +9,9 @@ import android.graphics.RectF
 import android.graphics.Typeface
 import android.util.AttributeSet
 import android.view.View
-import android.view.animation.Animation
-import android.view.animation.AnimationUtils
+import ru.musindev.myapp.R
 
 class RatingDonutView @JvmOverloads constructor(context: Context, attributeSet: AttributeSet? = null) : View(context, attributeSet) {
-
     //Овал для рисования сегментов прогресс бара
     private val oval = RectF()
     //Координаты центра View, а также Radius
@@ -30,10 +28,9 @@ class RatingDonutView @JvmOverloads constructor(context: Context, attributeSet: 
     private lateinit var strokePaint: Paint
     private lateinit var digitPaint: Paint
     private lateinit var circlePaint: Paint
-    private var inAnimation: Animation = AnimationUtils.loadAnimation(context, R.anim.alpha_in)
 
     init {
-        //Получаем атрибуты и устанавливаем их в соответствующие поля
+        //Получаем атрибуты и устанвливаем их в соответствующие поля
         val a =
             context.theme.obtainStyledAttributes(attributeSet, R.styleable.RatingDonutView, 0, 0)
         try {
@@ -43,101 +40,17 @@ class RatingDonutView @JvmOverloads constructor(context: Context, attributeSet: 
         } finally {
             a.recycle()
         }
-
         //Инициализируем первоначальные краски
         initPaint()
-    }
-
-    override fun onSizeChanged(w: Int, h: Int, oldw: Int, oldh: Int) {
-        radius = if (width > height) {
-            height.div(2f)
-        } else {
-            width.div(2f)
-        }
-    }
-
-    override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
-        val widthMode = MeasureSpec.getMode(widthMeasureSpec)
-        val widthSize = MeasureSpec.getSize(widthMeasureSpec)
-
-        val heightMode = MeasureSpec.getMode(heightMeasureSpec)
-        val heightSize = MeasureSpec.getSize(heightMeasureSpec)
-
-        val chosenWidth = chooseDimension(widthMode, widthSize)
-        val chosenHeight = chooseDimension(heightMode, heightSize)
-
-
-
-        val minSide = Math.min(chosenWidth, chosenHeight)
-        centerX = minSide.div(2f)
-        centerY = minSide.div(2f)
-
-        setMeasuredDimension(minSide, minSide)
-    }
-
-    private fun chooseDimension(mode: Int, size: Int) =
-        when (mode) {
-            MeasureSpec.AT_MOST, MeasureSpec.EXACTLY -> size
-            else -> 300
-        }
-
-    private fun drawRating(canvas: Canvas) {
-        //Здесь мы можем регулировать размер нашего кольца
-        val scale = radius * 0.8f
-        //Сохраняем канвас
-        canvas.save()
-        //Перемещаем нулевые координаты канваса в центр, вы помните, так проще рисовать все круглое
-        canvas.translate(centerX, centerY)
-        //Устанавливаем размеры под наш овал
-        oval.set(0f - scale, 0f - scale, scale , scale)
-        //Рисуем задний фон(Желательно его отрисовать один раз в bitmap, так как он статичный)
-        canvas.drawCircle(0f, 0f, radius, circlePaint)
-        //Рисуем "арки", из них и будет состоять наше кольцо + у нас тут специальный метод
-        canvas.drawArc(oval, -90f, convertProgressToDegrees(progress), false, strokePaint)
-        //Восстанавливаем канвас
-        canvas.restore()
-    }
-
-    fun setProgress(pr: Int) {
-        //Кладем новове значение в нашу поле класса
-        progress = pr
-        //Создаем краски с новыми цветами
-        initPaint()
-        //вызываем перерисовку View
-        invalidate()
-    }
-
-    private fun convertProgressToDegrees(progress: Int): Float = progress * 3.6f
-
-    @SuppressLint("DefaultLocale")
-    private fun drawText(canvas: Canvas) {
-        //Форматируем текст, чтобы мы выводили дробное число с одной цифрой после точки
-        val message = String.format("%.1f", progress / 10f)
-        //Получаем ширину и высоту текста, чтобы компенсировать их при отрисовке, чтобы текст был
-        //точно в центре
-        val widths = FloatArray(message.length)
-        digitPaint.getTextWidths(message, widths)
-        var advance = 0f
-        for (width in widths) advance += width
-        //Рисуем наш текст
-        canvas.drawText(message, centerX - advance / 2, centerY  + advance / 4, digitPaint)
-    }
-
-    override fun onDraw(canvas: Canvas) {
-        //Рисуем кольцо и задний фон
-        drawRating(canvas)
-        //Рисуем цифры
-        drawText(canvas)
     }
 
     private fun initPaint() {
         //Краска для колец
         strokePaint = Paint().apply {
             style = Paint.Style.STROKE
-
             //Сюда кладем значение из поля класса, потому как у нас краски будут видоизменяться
             strokeWidth = stroke
-            //Цвет мы тоже будем получать в специальном методе, потому что в зависимости от рейтинга
+            //Цвет мы тоже будем получать в специальном метода, потому что в зависимости от рейтинга
             //мы будем менять цвет нашего кольца
             color = getPaintColor(progress)
             isAntiAlias = true
@@ -159,6 +72,84 @@ class RatingDonutView @JvmOverloads constructor(context: Context, attributeSet: 
         }
     }
 
+    override fun onSizeChanged(w: Int, h: Int, oldw: Int, oldh: Int) {
+        radius = if (width > height) {
+            height.div(2f)
+        } else {
+            width.div(2f)
+        }
+    }
+
+    override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
+        val widthMode = MeasureSpec.getMode(widthMeasureSpec)
+        val widthSize = MeasureSpec.getSize(widthMeasureSpec)
+
+        val heightMode = MeasureSpec.getMode(heightMeasureSpec)
+        val heightSize = MeasureSpec.getSize(heightMeasureSpec)
+
+        val chosenWidth = chooseDimension(widthMode, widthSize)
+        val chosenHeight = chooseDimension(heightMode, heightSize)
+
+        val minSide = Math.min(chosenWidth, chosenHeight)
+        centerX = minSide.div(2f)
+        centerY = minSide.div(2f)
+
+        setMeasuredDimension(minSide, minSide)
+    }
+
+    private fun chooseDimension(mode: Int, size: Int) =
+        when (mode) {
+            MeasureSpec.AT_MOST, MeasureSpec.EXACTLY -> size
+            else -> 300
+        }
+
+    override fun onDraw(canvas: Canvas) {
+        //Рисуем кольцо и задний фон
+        drawRating(canvas)
+        //Рисуем цифры
+        drawText(canvas)
+    }
+
+    fun setProgress(pr: Int) {
+        //Кладем новове значение в нашу поле класса
+        progress = pr
+        //Создаем краски с новыми цветами
+        initPaint()
+        //вызываем перерисовку View
+        invalidate()
+    }
+
+    private fun drawRating(canvas: Canvas) {
+        //Здесь мы можем регулировать размер нашего кольца
+        val scale = radius * 0.8f
+        //Сохраняем канвас
+        canvas.save()
+        //Перемещаем нулевые координаты канваса в центр, вы помните, так проще рисовать все круглое
+        canvas.translate(centerX, centerY)
+        //Устанавливаем размеры под наш овал
+        oval.set(0f - scale, 0f - scale, scale , scale)
+        //Рисуем задний фон(Желательно его отрисовать один раз в bitmap, так как он статичный)
+        canvas.drawCircle(0f, 0f, radius, circlePaint)
+        //Рисуем "арки" из них и будет состоять наше кольцо + у нас тут специальный метод
+        canvas.drawArc(oval, -90f, convertProgressToDegrees(progress), false, strokePaint)
+        //Восстанавливаем канвас
+        canvas.restore()
+    }
+
+    @SuppressLint("DefaultLocale")
+    private fun drawText(canvas: Canvas) {
+        //Форматируем текст, чтобы мы выводили дробное число с одной цифрой после точки
+        val message = String.format("%.1f", progress / 10f)
+        //Получаем ширину и высоту текста, чтобы компенсировать их при отрисовке, что бы текст был
+        //точно в центре
+        val widths = FloatArray(message.length)
+        digitPaint.getTextWidths(message, widths)
+        var advance = 0f
+        for (width in widths) advance += width
+        //Рисуем наш текст
+        canvas.drawText(message, centerX - advance / 2, centerY  + advance / 4, digitPaint)
+    }
+
     private fun getPaintColor(progress: Int): Int = when(progress) {
         in 0 .. 25 -> Color.parseColor("#e84258")
         in 26 .. 50 -> Color.parseColor("#fd8060")
@@ -166,6 +157,6 @@ class RatingDonutView @JvmOverloads constructor(context: Context, attributeSet: 
         else -> Color.parseColor("#b0d8a4")
     }
 
-
-
+    private fun convertProgressToDegrees(progress: Int): Float = progress * 3.6f
 }
+
