+++ app/src/main/java/ru/musindev/myapp/view/fragments/FavoritesFragment.kt
@@ -1,4 +1,4 @@
-package ru.musindev.myapp
+package ru.musindev.myapp.view.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,6 +7,11 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.LinearLayoutManager
 import ru.musindev.myapp.databinding.FragmentFavoritesBinding
+import ru.musindev.myapp.domain.Film
+import ru.musindev.myapp.utils.AnimationHelper
+import ru.musindev.myapp.view.rv_adapters.FilmListRecyclerAdapter
+import ru.musindev.myapp.view.MainActivity
+import ru.musindev.myapp.view.rv_adapters.TopSpacingItemDecoration
 
 
 class FavoritesFragment : Fragment() {
