+++ app/src/main/java/ru/musindev/myapp/view/fragments/SelectionsFragment.kt
@@ -1,26 +1,27 @@
-package ru.musindev.myapp
+package ru.musindev.myapp.view.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import ru.musindev.myapp.databinding.FragmentWatchLaterBinding
+import ru.musindev.myapp.databinding.FragmentSelectionsBinding
+import ru.musindev.myapp.utils.AnimationHelper
 
-class WatchLaterFragment : Fragment() {
+class SelectionsFragment : Fragment() {
 
-    private lateinit var binding: FragmentWatchLaterBinding
+    private lateinit var binding: FragmentSelectionsBinding
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentWatchLaterBinding.inflate(inflater, container, false)
+        binding = FragmentSelectionsBinding.inflate(inflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        AnimationHelper.performFragmentCircularRevealAnimation(binding.watchLaterFragmentRoot, requireActivity(), 3)
+        AnimationHelper.performFragmentCircularRevealAnimation(binding.selectionsFragmentRoot, requireActivity(), 4)
     }
 }
