+++ app/src/main/java/ru/musindev/myapp/view/rv_viewholders/FilmViewHolder.kt
@@ -1,12 +1,13 @@
-package ru.musindev.myapp
+package ru.musindev.myapp.view.rv_viewholders
 
 import android.view.View
-import android.view.animation.Animation
-import android.view.animation.AnimationUtils
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
+import ru.musindev.myapp.R
+import ru.musindev.myapp.domain.Film
+import ru.musindev.myapp.view.customviews.RatingDonutView
 
 //В конструктор класс передается layout, который мы создали(film_item.xml)
 class FilmViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
