+++ app/src/main/java/ru/musindev/myapp/viewmodel/HomeFragmentViewModel.kt
@@ -1,0 +1,18 @@
+package ru.musindev.myapp.viewmodel
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import ru.musindev.myapp.App
+import ru.musindev.myapp.domain.Film
+import ru.musindev.myapp.domain.Interactor
+
+class HomeFragmentViewModel : ViewModel() {
+    val filmsListLiveData:  MutableLiveData<List<Film>> = MutableLiveData()
+    //Инициализируем интерактор
+    private var interactor: Interactor = App.instance.interactor
+
+    init {
+        val films = interactor.getFilmsDB()
+        filmsListLiveData.postValue(films)
+    }
+}
