+++ app/src/main/res/layout/film_item.xml
@@ -9,7 +9,7 @@
     android:elevation="1dp"
     app:cardCornerRadius="10dp">
 
-    <ru.musindev.myapp.RatingDonutView
+    <ru.musindev.myapp.view.customviews.RatingDonutView
         android:id="@+id/rating_donut"
         android:layout_width="50dp"
         android:layout_height="50dp"
