+++ app/src/main/res/layout/fragment_favorites.xml
@@ -6,7 +6,7 @@
     android:background="@drawable/background_favorites"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/favorites_fragment_root"
-    tools:context=".FavoritesFragment">
+    tools:context=".view.fragments.FavoritesFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/favorites_recycler"
