+++ app/src/main/res/layout/fragment_selections.xml
@@ -5,6 +5,6 @@
     android:layout_height="match_parent"
     android:id="@+id/selections_fragment_root"
     android:background="@drawable/background_podbor"
-    tools:context=".SelectionsFragment">
+    tools:context=".view.fragments.SelectionsFragment">
 
 </FrameLayout>
